+++ source/engine/makaba/vote/parseVoteResponse.js
@@ -4,15 +4,15 @@
  * @return {boolean} Returns `true` if the vote has been accepted. Returns `false` if the user has already voted for this thread or comment.
  */
 export default function parseVoteResponse(response) {
-	if (response.Error) {
-		// "Вы уже двачевали этот пост."
-		if (response.Error === -4) {
+	if (response.result === 1) {
+		return true
+	}
+	if (response.error) {
+		// If the user has already voted for the comment.
+		if (response.error.code === -4) {
 			return false
 		}
-		throw new Error(response.Reason)
+		throw new Error(response.error.message)
 	}
-	if (response.Status !== 'OK') {
-		throw new Error(response.Status)
-	}
-	return true
+	throw new Error()
 }